Support chiploads < 0.02 mm
As the chip thickness correction is interpolated from a table in the
Machienery's Handbook, we're limited to the values provided there.
Although the HB does not provide details about the origin of those
values, in literature one can find that the chip thickness correction
is a power function. Solving for the two table entries "0.3 -> 1" and
"1 -> 0.78" results in C = 0.78 * h^-0.2020636791...
Fitting the all of the table values results in C = 0.7834048 * h^-0.1971463
For simplicity I choose C = 0.78 * h^-0.2 -- all the table values are
rounded anyway and the difference is neglectible.